- Some macros replaced by functions - Mathtools: Added weighed_lower_bound - Formulaeditor can be resized -

+++ GeometryViewer.h
@@ -162,8 +162,8 @@ public:
   int dispNumTraj;  // displayed number of trajectory points
   #endif
   
-  int dispNumHits; // displayed number of lines and hits
-  int dispNumLeaks; // displayed number of leaks
+  size_t dispNumHits; // displayed number of lines and hits
+  size_t dispNumLeaks; // displayed number of leaks
   double transStep;  // translation step
   double angleStep;  // angle step
